The Devise documentation has a Lazy Way, that only works if the additional parameters have values from ActionController::StrongParameters list of permitted scalars e.g. String, Symbol, Date etc
If you have additional parameters that have values that are not from the list of permitted scalars, you need to let AC::SP know that a parameter has a different type e.g. for an array of scalar values, use the <field: []> notation (if you have nested hashes, you need to let AC::SP know you permit the key the hash resides in, and the permitted keys in that hash, and so on down for all levels of nesting.

See
* http://www.railsexperiments.com/using-strong-parameters-with-nested-forms/
* http://blog.sensible.io/2013/08/17/strong-parameters-by-example.html
* https://github.com/rails/strong_parameters#nested-parameters

If you have even more complex behaviour - say a dynamic set of attributes that can't be determined at design time, but need runtime construction, you will need the more heavy-weight full controller solution
See
* http://stackoverflow.com/questions/17767449/rails-4-0-with-devise-nested-attributes-unpermited-parameters